update delete cats

+++ admin/inc/class.bocategories.inc.php
@@ -143,9 +143,19 @@
 			}
 		}
 
-		function delete($cat_id,$drop_subs,$modify_subs)
+		function delete($data)
 		{
-			$this->cats->delete($cat_id,$drop_subs,$modify_subs);
+			if (is_array($data))
+			{
+				$cat_id			= $data['cat_id'];
+				$drop_subs		= (isset($data['drop_subs'])?$data['drop_subs']:False);
+				$modify_subs	= (isset($data['modify_subs'])?$data['modify_subs']:False);				
+
+				if ($cat_id > 0)
+				{
+					$this->cats->delete($cat_id,$drop_subs,$modify_subs);
+				}
+			}
 		}
 
 		function check_values($values)
